Creation of checks

+++ src/modules/curso/curso.service.ts
@@ -1,7 +1,8 @@
 import { Injectable } from '@nestjs/common';
-import { NotFoundException } from '@nestjs/common/exceptions';
+import { BadRequestException, NotFoundException } from '@nestjs/common/exceptions';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
+import { MatriculaService } from '../matricula/matricula.service';
 import { CreateCursoDto } from './dto/create-curso.dto';
 import { UpdateCursoDto } from './dto/update-curso.dto';
 import { Curso } from './entities/curso.entity';
@@ -11,14 +12,15 @@ export class CursoService {
   
   constructor(
     @InjectRepository(Curso)
-    private readonly cursoRepository: Repository<Curso>
+    private readonly cursoRepository: Repository<Curso>,
+    private readonly matriculaService: MatriculaService
   ) {}
 
   async create(createCursoDTO: CreateCursoDto) {
     const createdCurso = this.cursoRepository.create({
-      ...createCursoDTO
+      ...createCursoDTO,
     });
-    
+
     return await this.cursoRepository.save(createdCurso);
   }
 
@@ -37,9 +39,9 @@ export class CursoService {
   findAll() {
     return this.cursoRepository.find({
       relations: {
-        matriculas: true
+        matriculas: true,
       },
-      order: { id: 'ASC'}
+      order: { id: 'ASC' },
     });
   }
 
@@ -47,25 +49,28 @@ export class CursoService {
     const curso = await this.cursoRepository.findOne({
       where: { id },
       relations: {
-        matriculas: true
-      }
+        matriculas: true,
+      },
     });
 
     if (!curso)
       throw new NotFoundException(`O curso ${id} não foi encontrado.`);
-      
+
     return curso;
   }
-  
+
   async remove(id: number) {
     const curso = await this.cursoRepository.findOne({
       where: { id },
     });
+    
+    const matricula = this.matriculaService.findMatriculaByCurso(id);
 
+    if(matricula)
+      throw new BadRequestException("Este curso não pode ser removido pois há matrículas vinculadas.")
+      
     if (!curso)
       throw new NotFoundException(`O curso ${id} não foi encontrado.`);
-
-     
 
     return this.cursoRepository.remove(curso);
   }
